[FIX] Fix pre-commit errors
- [REF] Refactor CMakeLists.txt, update documentation, and improve code formatting across multiple files

+++ gravity_compensation_pd_controller/include/gravity_compensation_pd_controller/gravity_compensation_pd_controller.hpp
@@ -1,32 +1,37 @@
-/* -------------------------------------------------------------------
- *
- * This module has been developed by the Automatic Control Group
- * of the University of Salerno, Italy.
- *
- * Title:   gravity_compensation_pd_controller.hpp
- * Author:  Davide Risi
- * Org.:    UNISA
- * Date:    Sept 4, 2025
- *
- * This module implements a PD controller with gravity compensation.
- *
- * -------------------------------------------------------------------
- */
+// Copyright (c) 2025, University of Salerno, Automatic Control Group
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// Authors: Davide Risi
 
 #ifndef GRAVITY_COMPENSATION_PD_CONTROLLER__GRAVITY_COMPENSATION_PD_CONTROLLER_HPP_
 #define GRAVITY_COMPENSATION_PD_CONTROLLER__GRAVITY_COMPENSATION_PD_CONTROLLER_HPP_
 
-#include <Eigen/Core>
+#include "gravity_compensation_pd_controller/gravity_compensation_pd_controller_parameters.hpp"
 
+#include <inverse_dynamics_solver/inverse_dynamics_solver.h>
+
+#include <Eigen/Core>
+#include <memory>
+#include <vector>
+
+#include <controller_interface/chainable_controller_interface.hpp>
+#include <pluginlib/class_loader.hpp>
 #include <rclcpp/duration.hpp>
 #include <rclcpp/node_interfaces/node_parameters_interface.hpp>
 #include <rclcpp/time.hpp>
 #include <rclcpp_lifecycle/state.hpp>
-#include <pluginlib/class_loader.hpp>
-#include <controller_interface/chainable_controller_interface.hpp>
-#include <inverse_dynamics_solver/inverse_dynamics_solver.h>
 
-#include "gravity_compensation_pd_controller/gravity_compensation_pd_controller_parameters.hpp"
 namespace gravity_compensation_pd_controller
 {
 
@@ -41,13 +46,13 @@ struct RobotJointState
  * @brief A class representing a PD controller with gravity compensation.
  *
  * This class implements the \c controller_interface::ChainableControllerInterface interface.
- * All of the public methods override the corresponding methods of the \c controller_interface::ChainableControllerInterface class.
- * Please refer to the documentation of the base class for more details.
+ * All of the public methods override the corresponding methods of the \c
+ * controller_interface::ChainableControllerInterface class. Please refer to the documentation of
+ * the base class for more details.
  */
 class GravityCompensationPDController : public controller_interface::ChainableControllerInterface
 {
 public:
-
   GravityCompensationPDController();
 
   controller_interface::CallbackReturn on_init() override;
@@ -56,27 +61,36 @@ public:
 
   controller_interface::InterfaceConfiguration state_interface_configuration() const override;
 
-  controller_interface::CallbackReturn on_configure(const rclcpp_lifecycle::State& previous_state) override;
+  controller_interface::CallbackReturn on_configure(
+    const rclcpp_lifecycle::State & previous_state) override;
 
-  controller_interface::CallbackReturn on_activate(const rclcpp_lifecycle::State& previous_state) override;
+  controller_interface::CallbackReturn on_activate(
+    const rclcpp_lifecycle::State & previous_state) override;
 
-  controller_interface::CallbackReturn on_deactivate(const rclcpp_lifecycle::State& previous_state) override;
+  controller_interface::CallbackReturn on_deactivate(
+    const rclcpp_lifecycle::State & previous_state) override;
 
-  controller_interface::CallbackReturn on_error(const rclcpp_lifecycle::State& previous_state) override;
+  controller_interface::CallbackReturn on_error(
+    const rclcpp_lifecycle::State & previous_state) override;
 
 protected:
-  // The following methods are overridden from the base class. Refer to the base class documentation for details.
+  // The following methods are overridden from the base class. Refer to the base class documentation
+  // for details.
   std::vector<hardware_interface::CommandInterface> on_export_reference_interfaces() override;
-  controller_interface::return_type update_reference_from_subscribers(const rclcpp::Time& time, const rclcpp::Duration& period) override;
+  controller_interface::return_type update_reference_from_subscribers(
+    const rclcpp::Time & time, const rclcpp::Duration & period) override;
   bool on_set_chained_mode(bool chained_mode) override;
-  controller_interface::return_type update_and_write_commands(const rclcpp::Time& time, const rclcpp::Duration& period) override;
+  controller_interface::return_type update_and_write_commands(
+    const rclcpp::Time & time, const rclcpp::Duration & period) override;
 
   /**
-   * @brief Computes the joint effort command using a PD control law with optional gravity compensation.
+   * @brief Computes the joint effort command using a PD control law with optional gravity
+   * compensation.
    *
-   * This method calculates the joint effort command according to the control law described in the README file.
-   * If \c compensate_gravity parameter is disabled, the control law reduces to a standard PD controller.
-   * The resulting effort command is stored in the \c joint_command_ attribute.
+   * This method calculates the joint effort command according to the control law described in the
+   * README file. If \c compensate_gravity parameter is disabled, the control law reduces to a
+   * standard PD controller. The resulting effort command is stored in the \c joint_command_
+   * attribute.
    */
   void compute_control_law_();
 
@@ -124,7 +138,8 @@ protected:
   std::vector<double> joint_command_, joint_reference_, last_joint_reference_;
 
   /**
-   * @brief Shared pointer to the parameter listener responsible for handling the controller's parameters.
+   * @brief Shared pointer to the parameter listener responsible for handling the controller's
+   * parameters.
    */
   std::shared_ptr<gravity_compensation_pd_controller::ParamListener> parameter_handler_;
 
@@ -134,7 +149,8 @@ protected:
   Eigen::VectorXd torque_limits_;
 
   /**
-   * @brief Eigen vector to store the position error between the reference and current joint positions.
+   * @brief Eigen vector to store the position error between the reference and current joint
+   * positions.
    */
   Eigen::VectorXd position_error_;
 
@@ -154,20 +170,20 @@ protected:
   Eigen::DiagonalMatrix<double, Eigen::Dynamic> Kd_;
 
   /**
-   * @brief Constant to store the duration of the throttle interval as an integral value in milliseconds.
+   * @brief Constant to store the duration of the throttle interval as an integral value in
+   * milliseconds.
    */
-  static constexpr unsigned short DURATION_MS_{ 1000 };
+  static constexpr int DURATION_MS_{1000};
 
   /**
    * @brief Number of joints to control.
    */
-  std::size_t num_joints_{ 0 };
+  std::size_t num_joints_{0};
 
   /**
    * @brief Internal variable to store the current joint state of the robot.
    */
   RobotJointState robot_joint_state_;
-
 };
 
 }  // namespace gravity_compensation_pd_controller
